functionize the code
